requirements.txt
python-telegram-bot==20.6
gspread==6.1.2
oauth2client==4.1.3
pandas==2.2.2
apscheduler==3.10.4
openpyxl==3.1.2
python-dotenv==1.0.1
